requirements.txt
ccxt
datetime
requests
discord.py
